Extract scalar-value parsing into yaml-scalar.ts
Move stripQuotes, splitInlineComment, parseScalar, and parseFlowList out
of the already-oversized yaml-section-values.ts into a focused module,
so this PR doesn't grow that file further. Behavior-free.
