+++ src/util/yaml-section-values.ts
@@ -10,6 +10,12 @@
 import { ESPHOME_YAML_INDENT } from "./esphome-yaml-lang.js";
 import { LIST_SECTIONS } from "./section-entry-overrides.js";
 import {
+  parseFlowList,
+  parseScalar,
+  splitInlineComment,
+  stripQuotes,
+} from "./yaml-scalar.js";
+import {
   buildSplicedBody,
   yamlValueEqual,
   type KeySpan,
@@ -17,7 +23,6 @@ import {
 } from "./yaml-section-splice.js";
 import {
   formatYamlScalar,
-  parseYamlBoolean,
   serializeYamlValues,
   YamlRawValue,
   type SerializeYamlOptions,
@@ -291,76 +296,6 @@ const isChildListItemLine = (line: string, parentIndent: string): boolean => {
   if (lead.length < parentIndent.length) return false;
   const tail = line.slice(lead.length);
   return tail === "-" || tail.startsWith("- ");
-};
-
-const stripQuotes = (s: string): string => {
-  if ((s.startsWith('"') && s.endsWith('"')) || (s.startsWith("'") && s.endsWith("'"))) {
-    return s.slice(1, -1);
-  }
-  return s;
-};
-
-/**
- * Split a scalar's raw text into its value and a trailing inline
- * comment (``true #hides`` → ``{ value: "true", comment: " #hides" }``).
- * A ``#`` only starts a comment when it's whitespace-preceded and
- * outside quotes — ``Bedroom#2`` and ``"a # b"`` keep the ``#`` in the
- * value. ``comment`` retains its leading whitespace (``""`` when none)
- * so the serializer can re-append it verbatim.
- */
-const splitInlineComment = (raw: string): { value: string; comment: string } => {
-  let inSingle = false;
-  let inDouble = false;
-  for (let i = 0; i < raw.length; i++) {
-    const c = raw[i];
-    // Backslash escapes the next char inside a double-quoted scalar
-    // (`"a \" # b"`), so it can't desync the quote tracker. Single
-    // quotes escape via `''`, which the toggle already handles.
-    if (c === "\\" && inDouble) {
-      i++;
-      continue;
-    }
-    if (c === '"' && !inSingle) inDouble = !inDouble;
-    else if (c === "'" && !inDouble) inSingle = !inSingle;
-    else if (
-      c === "#" &&
-      !inSingle &&
-      !inDouble &&
-      (raw[i - 1] === " " || raw[i - 1] === "\t")
-    ) {
-      let ws = i;
-      while (ws > 0 && (raw[ws - 1] === " " || raw[ws - 1] === "\t")) ws--;
-      return { value: raw.slice(0, ws), comment: raw.slice(ws) };
-    }
-  }
-  return { value: raw, comment: "" };
-};
-
-// Quoting in YAML is the explicit "treat me as a string" signal —
-// ``key: "on"`` must stay the literal ``"on"`` even though ``on`` is
-// a truthy spelling. Detect the quotes BEFORE stripping so we only
-// run the boolean coercion on plain scalars; otherwise a string
-// field that happens to hold ``"on"`` / ``"yes"`` would silently
-// flip to boolean ``true`` on round-trip.
-const parseScalar = (raw: string): unknown => {
-  // Strip a trailing inline comment so a boolean/number field coerces
-  // and the form value isn't polluted with `# ...` text (#1235).
-  const { value: scalar } = splitInlineComment(raw);
-  const wasQuoted =
-    (scalar.startsWith('"') && scalar.endsWith('"')) ||
-    (scalar.startsWith("'") && scalar.endsWith("'"));
-  const v = stripQuotes(scalar);
-  if (!wasQuoted) {
-    const bool = parseYamlBoolean(v);
-    if (bool !== null) return bool;
-  }
-  return v;
-};
-
-const parseFlowList = (raw: string): string[] => {
-  const inner = raw.slice(1, -1).trim();
-  if (inner === "") return [];
-  return inner.split(",").map((p) => stripQuotes(p.trim()));
 };
 
 const collectBlockListItems = (
